Migrate existing tests to utils/test_beam_utils.py
from calc_vis/sample_jones_test.py
